added coli dependency, updated dart string maker logic to replace \r
