+++ .gitmodules
@@ -17,3 +17,6 @@
 [submodule "packages/reflect-core"]
 	path = packages/reflect-core
 	url = https://github.com/bridgedxyz/reflect-core-ts.git
+[submodule "packages/coli"]
+	path = packages/coli
+	url = https://github.com/bridgedxyz/CoLI.git
