+++ figma/src/flutter/make/index.ts
@@ -6,11 +6,14 @@ import { makeEdgeInsets } from "./edge-insets.make";
 import { makeShape } from "./shape.make";
 import { makeTextStyle } from "./text-style.make";
 import { makeText } from "./text.make";
+import { makeBoxDecoration, makeBoxDecorationBg } from "./box-decoration.make"
 
 export {
     makeBorderRadius,
     makeBorder,
     makeBoxShadow,
+    makeBoxDecoration,
+    makeBoxDecorationBg as makeBoxDecorationFill,
     makeColor,
     makeColorFromRGBO,
     makeEdgeInsets,
